# Avoid creating gridfunctions for the time derivatives of MFEM AuxVariables