--- a/framework/doc/content/source/mfem/kernels/MFEMKernel.md
+++ b/framework/doc/content/source/mfem/kernels/MFEMKernel.md
@@ -9,7 +9,7 @@ Base class for MFEM kernels applied to the weak form being solved.
 ## Overview
 
 MFEM kernels are responsible for providing domain integrators (inheriting from
-`mfem::BilinearFormIntegrator` or `mfem::LinearFormIntegrator` to add to the weak form of the FE
+`mfem::BilinearFormIntegrator` or `mfem::LinearFormIntegrator`) to add to the weak form of the FE
 problem accumulated in [EquationSystem](source/mfem/equation_systems/EquationSystem.md), along with any
 required marker arrays to restrict the integrator(s) to subdomains.