--- a/framework/include/mfem/mesh/MFEMMesh.h
+++ b/framework/include/mfem/mesh/MFEMMesh.h
@@ -9,15 +9,6 @@
 
 #ifdef MOOSE_MFEM_ENABLED
 
-//* This file is part of the MOOSE framework
-//* https://mooseframework.inl.gov
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
 #pragma once
 #include "FileMesh.h"
 #include "libmesh/ignore_warnings.h"