--- a/framework/src/mfem/mesh/MFEMMesh.C
+++ b/framework/src/mfem/mesh/MFEMMesh.C
@@ -9,15 +9,6 @@
 
 #ifdef MOOSE_MFEM_ENABLED
 
-//* This file is part of the MOOSE framework
-//* https://mooseframework.inl.gov
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
 #include "MFEMMesh.h"
 #include "libmesh/mesh_generation.h"