--- a/test/tests/mfem/ics/gold/OutputData/TransientScalarIC/Run0/Cycle000000/data.pvtu
+++ b/test/tests/mfem/ics/gold/OutputData/TransientScalarIC/Run0/Cycle000000/data.pvtu
@@ -11,7 +11,6 @@
 </PCells>
 <PPointData>
 <PDataArray type="Float64" Name="dh1_scalar_dt" NumberOfComponents="1"  format="ascii" />
-<PDataArray type="Float64" Name="dl2_scalar_dt" NumberOfComponents="1"  format="ascii" />
 <PDataArray type="Float64" Name="h1_scalar" NumberOfComponents="1"  format="ascii" />
 <PDataArray type="Float64" Name="l2_scalar" NumberOfComponents="1"  format="ascii" />
 </PPointData>