--- a/test/tests/mfem/variables/gold/OutputData/VariableSetupTest/Run0/Run0.pvd
+++ b/test/tests/mfem/variables/gold/OutputData/VariableSetupTest/Run0/Run0.pvd
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<VTKFile type="Collection" version="0.1" byte_order="LittleEndian">
+<VTKFile type="Collection" version="2.2" byte_order="LittleEndian">
 <Collection>
 <DataSet timestep="0" group="" part="0" file="Cycle000000/data.pvtu" name="mesh"/>
 <DataSet timestep="1" group="" part="0" file="Cycle000001/data.pvtu" name="mesh"/>