… test/tests/mfem/ics/gold/OutputData/TransientScalarIC/Run0/Cycle000000/proc000000.vtu
… test/tests/mfem/ics/gold/OutputData/TransientScalarIC/Run0/Cycle000001/proc000000.vtu