## Reason
We do not currently populate the gridfunctions for the time derivatives of MFEM AuxVariables, nor do I think we should.

## Design
Split the code paths for the creation of MFEM AuxVariables and MFEM Variables.

## Impact
Less confusing outputs (the time derivatives are effectively wrong atm) and faster xmldiffs.

Note `MFEMTransientScalarIC` uses a very small mesh, and the difference get bigger for bigger meshes (my real motivation is the new test in #31164 - in that case the new time is definitely not proportional, it's _much_ faster for some reason...).
Before (total time suggests ~1.7s to compare four vars):
```
./run_tests -C tests/mfem/ics/ --re MFEMTransientScalarIC
Parsing application capabilities... 1.27 seconds
mfem/ics.MFEMTransientScalarIC ........................................................................................... OK

-----------------------------------------------------------------------------------------------------------------------------
Ran 1 tests in 3.3 seconds. Average test time 1.6 seconds, maximum test time 1.6 seconds.
1 passed, 0 skipped, 0 failed
```

After (total time suggests ~1.3s to compare three vars, which is actually proportional to the above):
```
./run_tests -C tests/mfem/ics/ --re MFEMTransientScalarIC
Parsing application capabilities... 1.31 seconds
mfem/ics.MFEMTransientScalarIC ........................................................................................... OK

-----------------------------------------------------------------------------------------------------------------------------
Ran 1 tests in 2.9 seconds. Average test time 1.6 seconds, maximum test time 1.6 seconds.
1 passed, 0 skipped, 0 failed
```